lambda_func/requirements.txt
numpy==1.14.5
opencv-python==3.4.2.16
Pillow
